Update revert array
Write a program that takes an array of size N as input and gives the output as an array in the reverse order. The format of the input is as follow:
N
a1 a2 a3 a4 a5...an
N is the size of the array and a1,a2,a3,...an, are its elements. Your program should give output as follows:
an an-1,an-2...a1